ocrd_utils: group utility functions in modules
